third_party/blink/web_tests/platform/linux-chrome/external/wpt/picture-in-picture/idlharness.window-expected.txt

This is a testharness.js-based test.
Found 9 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] idl_test setup
  promise_test: Unhandled rejection with value: object "NotSupportedError: Picture-in-Picture is not available."
[FAIL] PictureInPictureWindow must be primary interface of pipw
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: pipw is not defined"
[FAIL] Stringification of pipw
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: pipw is not defined"
[FAIL] PictureInPictureWindow interface: pipw must inherit property "width" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: pipw is not defined"
[FAIL] PictureInPictureWindow interface: pipw must inherit property "height" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: pipw is not defined"
[FAIL] PictureInPictureWindow interface: pipw must inherit property "onresize" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: pipw is not defined"
[FAIL] PictureInPictureEvent must be primary interface of new PictureInPictureEvent("type", { pictureInPictureWindow: pipw })
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: pipw is not defined"
[FAIL] Stringification of new PictureInPictureEvent("type", { pictureInPictureWindow: pipw })
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: pipw is not defined"
[FAIL] PictureInPictureEvent interface: new PictureInPictureEvent("type", { pictureInPictureWindow: pipw }) must inherit property "pictureInPictureWindow" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: pipw is not defined"
Harness: the test ran to completion.
